Add a shell.nix for people using nixOS
This is by no means a complete environment, but if no one minds me
leaving this here it would save me some trouble - but is not required.

If you are unfamiliar with nixos, this file is intended to give my
environment the necessary mongo utils / yarn / node and so forth.

You can peg a node version and all sorts of various magic. I would
recommend any nerds with sufficient free time investigate - but you will
find it very hard to escape once it has ensnared you.