Area ancestry deconvolution
There is a new, simple, source of truth for area structure derivations.
Each area document holds a parent reference, which points to its desired
reference (or none, if it is a root node). The derived fields such as

 - Children
 - Ancestry

have been collated in `embeddedRelations`. the denormalized fields
`ancestors` and `pathTokens` have also been colocated into a single
entity path that makes propogating updates to denormalized data easier
(e.g, when updating an areas name). It leaves us a space for adding new
fields, and finally exposes the ACTUAL id of the document so that we can
index effectively.

I would suggest that using the parent field is the most straightforward
and safest thing to do in pretty much all scenarios, but if you are a
mongodb veteran you may find need for the array indexing features -
though as far as I can tell, this is rarely the case?

 - Added tests
 - Updated some existing tests
 - Data returned at the gql layer should be identical, but I have not
   fully de-risked those tests. All I can say is that they are passing.

This commit does not address #126, but gives us an enviable headstart
at it. It is however needed by #407