write out all fail() invocations
I believe the fail() pattern came from jasmine down to jest in the first
instance, though I am not sure. In any event it leads to some strange
testing behavior in which you begin to re-invent the framework from the
inside out.

For this reason it seems the vite team has decided not to make a pass at
including it, since it is technically breaking containment even in the
context of jest testing.

I have done my best to take the time it needs to ascertain which of
these fails() should be asserts, and which should be using the standard
expectations api.