Passing most tests stable with vitest fixtures
The test speed is up roughly 10x but my expectation is that it should be
able to come lower than that. The beauty of the way test containment is
implemented here means that there is not much need for isolation - which
is good for simulating our use case because we make use of a lot of
singletons (though thankfully not TOO much shared state)

It was all relatively trivial with the exception of histories. I am
intensely anxious with the implementation that has been used for
document history, as it obscures control flow and does not look like it
will scale - especially if we were ever to want horizontal deployment.

All in all the tests are faster, and can tighten up development feedback
loops BUT the un-strict nature of our existing test runners means that
there are some inconsistencies in execution that I haven't yet uncovered